Relocate organisms folder to scr
