+++ src/organisms/footer/footer-section2/StyledFooterSection2.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { theme } from "../../../../themes/MainTheme";
+import { theme } from "../../../themes/MainTheme";
 
 export const StyledFooterSection2 = styled.div`
   width: 100%;
