+++ src/organisms/footer/footer-section2/index.ts
