+++ src/organisms/footer/footer_main_view/Footer.tsx
@@ -4,7 +4,7 @@ import { StyledFooterWrapper } from "./StyledFooter";
 import sideImage from "../../../assets/images/footer/-g-Group 73.svg";
 =======
 import sideImage from "../../../../assets/images/footer/-g-Group 73.svg";
-import { copyrights } from "../../../../assets/content/footer/SectionAbout";
+import { copyrights } from "../../../assets/content/footer/SectionAbout";
 import { FooterSection1 } from "../footer-section1";
 import { FooterSection2 } from "../footer-section2";
 import { FooterSection3 } from "../footer-section3";
