+++ src/organisms/footer/footer_main_view/index.ts
